craft-rest-api/requirements.txt
Flask==0.12
requests
py-healthcheck